move connect button to status bar
